requirements.txt
amqp==2.5.0
anyjson==0.3.3
appdirs==1.4.3
APScheduler==3.6.0
attrs==19.1.0
billiard==3.6.0.0
cached-property==1.5.1
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.2
django-appconf==1.0.3
django-apscheduler==0.3.0
django-jalali==3.1.0
django-jalali-date==0.2.9
django-timezone-field==3.0
django-watermark==0.1.8
djangorestframework==3.9.4
idna==2.8
isodate==0.6.0
jdatetime==3.2.0
lxml==4.3.4
Pillow==6.0.0
psycopg2==2.8.3
python-crontab==2.3.8
python-dateutil==2.8.0
pytz==2019.1
redis==3.2.1
requests==2.22.0
requests-toolbelt==0.9.1
schedule==0.6.0
six==1.12.0
sqlparse==0.3.0
tzlocal==1.5.1
urllib3==1.25.3
vine==1.3.0
zeep==3.4.0
